Integration test for DELETE-Request

+++ backend/src/main/java/de/neuefische/backend/product/ProductController.java
@@ -14,6 +14,7 @@ import java.util.List;
 public class ProductController {
 
     private final ProductService productService;
+    private final ProductRepository productRepository;
 
 
     @GetMapping
@@ -32,8 +33,12 @@ public class ProductController {
     }
 
     @DeleteMapping("{id}")
-    public void deleteProduct(@PathVariable String id) {
-        productService.deleteProduct(id);
+    public ResponseEntity<Object> deleteProduct(@PathVariable String id) {
+        if (productRepository.existsById(id)) {
+            productService.deleteProduct(id);
+            return ResponseEntity.noContent().build();
+        }
+        else return ResponseEntity.notFound().build();
     }
 
 }
